Refactor : ExtraHome 사용자 홈화면 상단 top과 Content 분리
홈화면 TopBar와 Content 한 컴포넌트 화면에 있던 것을 분리해 렌더링 최적화하였다.
