+++ src/testFlag.ts
@@ -1,2 +1,3 @@
-export { TEST_FLAG };
+export { TEST_FLAG, USER_NAME };
 const TEST_FLAG = true;
+const USER_NAME = "USER_NAME";
